Add a strategy utils module
This module will contain any utilities needed by the strategy functions. They are placed in a separate module to enable us to test all of the strategies by iterating over the dictionary of callables in the strategies module.

The only utility for the moment is the valid_history module which ensures that history strings passed to the strategies are valid ie. contain only c or d.