Change to nose testing for generator support
Since we want to automatically grab all of the strategies from the strategies module and run them through the same set of tests, we need a way to automatically generate tests. Unittest wasn't allowing me to do this and we getting really messy and ugly.

Nose makes this easy by allowing you to define tests that yield other functions which contain testing logic.