Remove the strategyutils module
This module is no longer needed since we're going to be moving to a model where each strategy is only fed one move at a time. As such, it's easy to check whether a move is valid. We're also able to farm out move validation to a method in the abstract base class, to be implemented in future commits