Rewrite tests for new class based strats
This implements new tests for the new class based strategies.

The only test currently implemented is a test that ensures that regardless of opponent history, strategy history and characterset, the strategies always return values from the characterset.

In order to achieve this we define a custom hypothesis strategy that generates a two element characterset, and then generates two random length text snippets of the same length (one for the history, one for the opponent history)

We then throw this hypothesis strategy at each strategy. Unfortunately it appears that at this time, hypothesis doesn't support nose's yield statement, which would allow for slightly nicer syntax in this test since we're testing the same thing with multiple data.