Update method for simulating normalised payoff
This updates the method simulate_expected_payoff in two ways.

Firstly, it updates it to simulate_normalised_payoff .ie we normalise the payoff we are calculating by 1-continuation_probability.

Secondly, it updates it to use the new class based strategies.

Finally, it now simultaneously calculates the normalised payoffs for both strategies.