Add classes to model payoff matrices
This adds a generic class which can be used to model all two player, two action payoff matrices, with a function to get the payoff given different actions.

This also adds a more specific PrisonersDilemmaPayoff class derived from the PayoffMatrix class, which allows users to define the payoff matrix in terms of the familiar P, R, S and T variables.