Add some tests for the calculations module
This adds some tests for the calculations module. Specifically we will be testing it against the theoretical exact values given in the literature.

This adds the first set of tests which covers the combinations of strategies that give only a single result for the entire game.

This also adds a few helper functions that will be used to verify more complex interactions of strategies.