Fix robustness of tolerance checking
This fixes a few issues with the tolerance checking method we have.

Firstly, it makes the tolerance relative to the size of the expected and actual values. This means that larger numbers can have larger errors.

It also just passes on infinite results, because we probably don't care about these.