Set timeout to infinite before running tests
This sets the default settings object for Hypothesis to have infinite timeout. I think tests are sporadically failing (test_calculations) or constantly failing (test_simulations) because the large length of the list of combinations coupled with the long computation time is causing them to timeout.

This method should be called once by nose prior to running any tests in theory, because it's part of the init of the tests module.