Add code for naive method for payoff with mistakes
This adds the implementation of a naive way of calculating an iterated sum for the payoff of an iterated prisoners dilemma with mistakes.

This is the exhaustive method with no optimisations added. Each term is calculated, and then the four resulting terms are pushed onto a queue. When coefficients become too small, they are no longer pushed. Thus, the queue should eventually empty at which point the totals are returned.