Fix value of coefficient being added to the queue
We need to only add the coefficient to the queue. However we want to only add coefficients to the queue if the _term_ is smaller than epsilon. In order to accomplish this in the easiest way, we have to leave the max value of the payoff matrix out of the coefficient variable, but then multiply before comparing.