Add multiprocessed version of monte-carlo
This adds a version of the monte-carlo simulation that takes advantage of multiprocessing.

This is implemented by splitting the number of trials into chunks, and sending these chunks to the separate processors.

In the case that the estimator standard deviation is required to be with a certain range, the method performs however many trials were given (default 1000) and then performs additional trials in batches until the correct stdev is achieved. This is necessary because performing individual additional trials has such high overhead that it negates any multiprocessing performance benefits.

There is currently a constant that determines how many extra trials are performed per batch in order to get closer to the estimator stdev. This is probably not the smartest way to do things. We could probably do something like adjust the the number of trials down as we get closer.