Processes keep track of totals interally
Constantly publishing to the result queue was leading to way too much overhead.

Now, processes just keep track of their own total for terms they've computed, and then when the queue is empty they publish this total to the result queue.

This results in a 70s time saving and many less function calls