Fix another str or list issue
This fixes a long standing issue (and what I suspect to be the cause of most of the random Travis build failures) where an empty history will give false to history == [] because it is in fact an empty string. This means that code is executed that expects a non-empty history, when in fact the history is empty.

This should fix this by instead comparing the length to zero. This should work for both strings and lists, because I'm fairly sure we're still using both.